requirements.txt
PySide6>=6.7.0
openai-whisper>=20250625
torch>=2.3.0
torchaudio>=2.3.0
ffmpeg-python>=0.2.0
